backend/pyproject.toml
[tool.poetry]
name = "env360-backend"
version = "1.0.0"
description = "Production-grade FastAPI backend with GraphQL, OAuth/SSO, and PostgreSQL"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
# FastAPI and server
fastapi = "0.115.0"
uvicorn = {extras = ["standard"], version = "0.32.0"}
python-multipart = "0.0.12"

# Database - PostgreSQL
sqlalchemy = {extras = ["asyncio"], version = ">=2.0.36"}
asyncpg = "0.30.0"
alembic = "1.14.0"
psycopg2-binary = "2.9.10"

# Authentication & OAuth
python-jose = {extras = ["cryptography"], version = "3.3.0"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
python-oauth2 = "1.1.1"
authlib = "1.3.1"
httpx = "0.27.2"

# Environment & Configuration
pydantic = "2.9.2"
pydantic-settings = "2.5.2"
python-dotenv = "1.0.1"
email-validator = "^2.0.0"

# GraphQL
strawberry-graphql = "0.217.0"

# Utilities
python-dateutil = ">=2.9.0.post0"
pytz = "2024.2"
dbos = "^2.10.0"
cryptography = "^46.0.4"
kubernetes = "^35.0.0"

[tool.poetry.group.dev.dependencies]
# Development tools
black = "24.10.0"
flake8 = "7.1.1"
mypy = "1.11.2"

[tool.poetry.group.test.dependencies]
# Testing
pytest = "8.3.3"
pytest-asyncio = "0.24.0"

[tool.poetry.scripts]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
